refactor: config: split clap parsing and RuntimeConfig into separate files
`config.rs` mixed three concerns: the `Config` struct itself, the
`RuntimeConfig` runtime-mutable subset, and clap-driven CLI parsing
(`parse_bytes_with_unit`, `parse_snapshot_policy`, `Config::build`).
Splitting them per the project's "one main type per file" rule isolates
the clap surface as a single feature-gated module rather than scattering
`#[cfg(feature = "clap")]` gates through the file, and moves
`RuntimeConfig` to its own file.

The `crate::config::RuntimeConfig` re-export is preserved via `mod.rs`,
so the three call sites (`raft_core.rs`, `raft_inner.rs`,
`raft/mod.rs`) need no edits.
