+++ openraft/src/config/runtime_config.rs
@@ -1,0 +1,20 @@
+//! Updatable config for a raft runtime.
+
+use std::sync::atomic::AtomicBool;
+
+use crate::Config;
+
+/// Updatable config for a raft runtime.
+pub(crate) struct RuntimeConfig {
+    pub(crate) enable_heartbeat: AtomicBool,
+    pub(crate) enable_elect: AtomicBool,
+}
+
+impl RuntimeConfig {
+    pub(crate) fn new(config: &Config) -> Self {
+        Self {
+            enable_heartbeat: AtomicBool::from(config.enable_heartbeat),
+            enable_elect: AtomicBool::from(config.enable_elect),
+        }
+    }
+}
